Create carry tracker component
